requirements.txt
requests==2.19.1
RPi.GPIO==0.6.3